nextgen_platform/apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
redis>=5.0.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
celery>=5.4.0
prometheus-fastapi-instrumentator>=7.0.0
prometheus-client>=0.20.0
email-validator>=2.2.0
alembic>=1.13.0
python-multipart>=0.0.9
